`TestCase`をトレイトオブジェクトにする
