--- crates/voicevox_core_c_api/tests/e2e/testcases.rs
@@ -1,1 +1,0 @@
-pub(crate) mod compatible_engine;
